feat: update n8n-users-list to no longer use preset list of actions
