+++ packages/design-system/src/components/N8nUsersList/UsersList.stories.ts
@@ -1,6 +1,7 @@
 import N8nUsersList from './UsersList.vue';
 import { action } from '@storybook/addon-actions';
 import type { StoryFn } from '@storybook/vue';
+import { IUser } from '@/types';
 
 export default {
 	title: 'Modules/UsersList',
@@ -21,12 +22,24 @@ const Template: StoryFn = (args, { argTypes }) => ({
 	components: {
 		N8nUsersList,
 	},
-	template: '<n8n-users-list v-bind="$props" @reinvite="onReinvite" @delete="onDelete" />',
+	template:
+		'<n8n-users-list v-bind="$props" :actions="actions" @reinvite="onReinvite" @delete="onDelete" />',
 	methods,
 });
 
 export const UsersList = Template.bind({});
 UsersList.args = {
+	actions: [
+		{
+			label: 'Resend Invite',
+			value: 'reinvite',
+			guard: (user: IUser) => !user.firstName,
+		},
+		{
+			label: 'Delete User',
+			value: 'delete',
+		},
+	],
 	users: [
 		{
 			id: '1',
