+++ packages/design-system/src/components/N8nUsersList/UsersList.vue
@@ -25,19 +25,13 @@
 </template>
 
 <script lang="ts">
-import { IUser } from '../../types';
+import { IUser, IUserListAction } from '../../types';
 import N8nActionToggle from '../N8nActionToggle';
 import N8nBadge from '../N8nBadge';
 import N8nUserInfo from '../N8nUserInfo';
 import Locale from '../../mixins/locale';
 import mixins from 'vue-typed-mixins';
-import { t } from '../../locale';
 import { PropType } from 'vue';
-
-export interface IUserListAction {
-	label: string;
-	value: string;
-}
 
 export default mixins(Locale).extend({
 	name: 'n8n-users-list',
@@ -61,17 +55,9 @@ export default mixins(Locale).extend({
 		currentUserId: {
 			type: String,
 		},
-		deleteLabel: {
-			type: String,
-			default: () => t('nds.usersList.deleteUser'),
-		},
-		reinviteLabel: {
-			type: String,
-			default: () => t('nds.usersList.reinviteUser'),
-		},
 		actions: {
-			type: Array as PropType<string[]>,
-			default: () => ['delete', 'reinvite'],
+			type: Array as PropType<IUserListAction[]>,
+			default: () => [],
 		},
 	},
 	computed: {
@@ -115,37 +101,16 @@ export default mixins(Locale).extend({
 	},
 	methods: {
 		getActions(user: IUser): IUserListAction[] {
-			const actions = [];
-			const DELETE: IUserListAction = {
-				label: this.deleteLabel,
-				value: 'delete',
-			};
-
-			const REINVITE: IUserListAction = {
-				label: this.reinviteLabel,
-				value: 'reinvite',
-			};
-
 			if (user.isOwner) {
 				return [];
 			}
 
-			if (!user.firstName) {
-				if (this.actions.includes('reinvite')) {
-					actions.push(REINVITE);
-				}
-			}
+			const defaultGuard = () => true;
 
-			if (this.actions.includes('delete')) {
-				actions.push(DELETE);
-			}
-
-			return actions;
+			return this.actions.filter((action) => (action.guard || defaultGuard)(user));
 		},
 		onUserAction(user: IUser, action: string): void {
-			if (action === 'delete' || action === 'reinvite') {
-				this.$emit(action, user.id);
-			}
+			this.$emit(action, user.id);
 		},
 	},
 });
